+++ src/app/src/mainwindow.cpp
@@ -575,6 +575,25 @@ bool MainWindow::stopLiveView_PostProcessing() {
 }
 
 bool MainWindow::startAcquisition() {
+    if (m_plateFormatCurrentIndex == -1) {
+        QMessageBox messageBox;
+        messageBox.setWindowTitle("Warning!");
+        messageBox.setText("Automatic analysis will not run because no plate format has been selected for this acquisition.\n"
+            "Click cancel and select the plate format to enable automatic analysis."
+            " To continue the acquisition with automatic analysis disabled, click continue.");
+        messageBox.setIcon(QMessageBox::NoIcon);
+        messageBox.addButton(QMessageBox::Cancel);
+        messageBox.addButton(QString("Continue"), QMessageBox::NoRole);
+
+        m_userCanceledAcquisition = messageBox.exec() == QMessageBox::Cancel;
+        if (m_userCanceledAcquisition) {
+            spdlog::info("User canceled acquisition start because no platemap was selected.");
+            return false;
+        } else {
+            spdlog::info("User selected to continue acquisition with no platemap selected.");
+        }
+    }
+
     spdlog::info("Starting acquisition");
 
     m_acquisitionThread = QThread::create(MainWindow::acquisitionThread, this);